Move test to APIUtil
