+++ features/apimgt/org.wso2.carbon.apimgt.keymanager.feature/pom.xml
@@ -169,7 +169,7 @@
                         </configuration>
                     </execution>
                     <execution>
-                        <id>copy</id>
+                        <id>copy-revokedjwt-webapp</id>
                         <phase>package</phase>
                         <goals>
                             <goal>copy</goal>
